Final Commit
BookTest class wasn't compiling prior to this edit for some reason, and I changed a comment in the code as you can see by the edits, and it now magically works. If there is any explanation to this or if you would like more information on my struggles (It's not a coding problem, its more of a being unfamiliar with BlueJ problem), I would be happy to talk about it via email/Office hours.